Some more Double -> Float

+++ korgw/src/macosMain/kotlin/korlibs/render/DefaultGameWindowMacos.kt
@@ -115,8 +115,8 @@ class MyNSOpenGLView(
         defaultGameWindow.dispatch(gestureEvent.also {
             it.type = GestureEvent.Type.SWIPE
             it.id = 0
-            it.amountX = event.deltaX.toDouble()
-            it.amountY = event.deltaY.toDouble()
+            it.amountX = event.deltaX.toFloat()
+            it.amountY = event.deltaY.toFloat()
         })
         //println("swipeWithEvent:event=$event")
         super.swipeWithEvent(event)
@@ -126,7 +126,7 @@ class MyNSOpenGLView(
         defaultGameWindow.dispatch(gestureEvent.also {
             it.type = GestureEvent.Type.SMART_MAGNIFY
             it.id = 0
-            it.amount = 1.0
+            it.amount = 1f
         })
         //println("smartMagnifyWithEvent:event=$event")
         super.smartMagnifyWithEvent(event)
@@ -174,7 +174,7 @@ class MyNSOpenGLView(
                 MouseEvent.Type.SCROLL -> 0
                 else -> buttonMask(e.buttonMask.toInt())
             },
-            scrollDeltaX = -e.deltaX, scrollDeltaY = -e.deltaY, scrollDeltaZ = -e.deltaZ,
+            scrollDeltaX = -e.deltaX.toFloat(), scrollDeltaY = -e.deltaY.toFloat(), scrollDeltaZ = -e.deltaZ.toFloat(),
             isShiftDown = e.shift, isCtrlDown = e.ctrl, isAltDown = e.alt, isMetaDown = e.meta,
             scrollDeltaMode = MouseEvent.ScrollDeltaMode.PIXEL
         )
